Server/requirements.txt
pika~=1.3.0
python-dotenv~=0.20.0
readchar~=3.0.5

